pyproject.toml
[project]
name = "supu-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cupy-cuda12x>=13.5.1",
    "requests>=2.32.4",
]
